Aplicado design patterns (service) para validações
