+++ src/app/controllers/SessionController.js
@@ -2,24 +2,8 @@ import User from '../models/User';
 import jwt from "jsonwebtoken";
 import bcrypt from "bcryptjs";
 import authConfig  from '../../config/auth'
-import * as Yup from "yup";
 class SessionController{
-
-
-    
     async logarUsuario(req,res){  
-        const schema = Yup.object().shape({
-            email: Yup.string()
-            .email()
-            .required(),
-            senha: Yup.string()
-            .required()
-        });
-
-        if (!(await schema.isValid(req.body))) {
-            return res.status(400).json({ error: "Erro de validação" });
-        }      
-  
         const { email,senha } = req.body;
         const user = await User.findOne({email:email});
